test(scheduler): split scheduler test responsibilities
Move the scheduler tests into focused API, discovery, execution, runtime, types, and worker files without changing test behavior or coverage.
