+++ internal/scheduler/scheduler_api_test.go
@@ -1,0 +1,96 @@
+package scheduler
+
+import (
+	"errors"
+	"slices"
+	"testing"
+
+	"github.com/docker/compose/v5/pkg/api"
+
+	"github.com/kimdre/doco-cd/internal/docker"
+)
+
+func TestSchedulerModes(t *testing.T) {
+	t.Parallel()
+
+	if got := schedulerModes(false); !slices.Equal(got, []scheduledJobMode{scheduledJobModeContainer}) {
+		t.Fatalf("schedulerModes(false) = %v, want compose only", got)
+	}
+
+	if got := schedulerModes(true); !slices.Equal(got, []scheduledJobMode{scheduledJobModeContainer, scheduledJobModeSwarm}) {
+		t.Fatalf("schedulerModes(true) = %v, want compose and swarm", got)
+	}
+}
+func TestFindRunnableJob(t *testing.T) {
+	t.Parallel()
+
+	validLabels := map[string]string{
+		docker.DocoCDJobLabels.JobEnabled:  "true",
+		docker.DocoCDJobLabels.JobSchedule: "@every 1m",
+	}
+
+	tests := []struct {
+		name      string
+		jobs      []scheduledJob
+		jobName   string
+		stackName string
+		wantErr   error
+	}{
+		{
+			name: "single matching job",
+			jobs: []scheduledJob{
+				{name: "stack-backup-1", labels: validLabels},
+			},
+			jobName: "stack-backup-1",
+		},
+		{
+			name: "stack filter avoids ambiguity",
+			jobs: []scheduledJob{
+				{name: "backup", labels: map[string]string{docker.DocoCDJobLabels.JobEnabled: "true", docker.DocoCDJobLabels.JobSchedule: "@every 1m", api.ProjectLabel: "stack-a"}},
+				{name: "backup", labels: map[string]string{docker.DocoCDJobLabels.JobEnabled: "true", docker.DocoCDJobLabels.JobSchedule: "@every 1m", api.ProjectLabel: "stack-b"}},
+			},
+			jobName:   "backup",
+			stackName: "stack-a",
+		},
+		{
+			name: "job not found",
+			jobs: []scheduledJob{
+				{name: "other", labels: validLabels},
+			},
+			jobName: "backup",
+			wantErr: ErrScheduledJobNotFound,
+		},
+		{
+			name: "job disabled",
+			jobs: []scheduledJob{
+				{name: "backup", labels: map[string]string{docker.DocoCDJobLabels.JobEnabled: "false", docker.DocoCDJobLabels.JobSchedule: "@every 1m"}},
+			},
+			jobName: "backup",
+			wantErr: ErrScheduledJobDisabled,
+		},
+		{
+			name: "ambiguous job name",
+			jobs: []scheduledJob{
+				{name: "backup", labels: validLabels},
+				{name: "backup", labels: validLabels},
+			},
+			jobName: "backup",
+			wantErr: ErrScheduledJobAmbiguous,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			t.Parallel()
+
+			_, _, err := findRunnableJob(tt.jobs, tt.jobName, tt.stackName)
+			if tt.wantErr == nil && err != nil {
+				t.Fatalf("findRunnableJob() unexpected error = %v", err)
+			}
+
+			if tt.wantErr != nil && !errors.Is(err, tt.wantErr) {
+				t.Fatalf("findRunnableJob() error = %v, want %v", err, tt.wantErr)
+			}
+		})
+	}
+}
