+++ internal/scheduler/scheduler_runtime_test.go
@@ -1,0 +1,221 @@
+package scheduler
+
+import (
+	"fmt"
+	"testing"
+	"time"
+
+	"github.com/kimdre/doco-cd/internal/docker"
+)
+
+func TestStatusForScheduledJob(t *testing.T) {
+	t.Parallel()
+
+	tests := []struct {
+		name          string
+		job           scheduledJob
+		cfg           docker.JobScheduleConfig
+		runtimeStatus string
+		running       bool
+		want          string
+	}{
+		{
+			name: "container one_off created without runtime status stays created",
+			job: scheduledJob{
+				mode:           scheduledJobModeContainer,
+				containerState: "created",
+			},
+			cfg:  docker.JobScheduleConfig{ExecutionMode: docker.JobExecutionModeOneOff},
+			want: "created",
+		},
+		{
+			name: "container one_off created with runtime status uses exit code",
+			job: scheduledJob{
+				mode:           scheduledJobModeContainer,
+				containerState: "created",
+			},
+			cfg:           docker.JobScheduleConfig{ExecutionMode: docker.JobExecutionModeOneOff},
+			runtimeStatus: "exited (143)",
+			want:          "exited (143)",
+		},
+		{
+			name: "container restart keeps docker state",
+			job: scheduledJob{
+				mode:            scheduledJobModeContainer,
+				containerState:  "exited",
+				containerStatus: "Exited (0) 2 seconds ago",
+			},
+			cfg:  docker.JobScheduleConfig{ExecutionMode: docker.JobExecutionModeRestart},
+			want: "exited (0)",
+		},
+		{
+			name: "swarm one_off not rewritten",
+			job: scheduledJob{
+				mode: scheduledJobModeSwarm,
+			},
+			cfg:           docker.JobScheduleConfig{ExecutionMode: docker.JobExecutionModeOneOff},
+			runtimeStatus: "exited (0)",
+			want:          "",
+		},
+		{
+			name: "running state has priority",
+			job: scheduledJob{
+				mode:           scheduledJobModeContainer,
+				containerState: "created",
+			},
+			cfg:           docker.JobScheduleConfig{ExecutionMode: docker.JobExecutionModeOneOff},
+			runtimeStatus: "exited (0)",
+			running:       true,
+			want:          "running",
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			t.Parallel()
+
+			got := statusForScheduledJob(tt.job, tt.cfg, tt.runtimeStatus, tt.running)
+			if got != tt.want {
+				t.Fatalf("statusForScheduledJob()=%q want=%q", got, tt.want)
+			}
+		})
+	}
+}
+
+// TestSetRuntimeStatesSnapshot_PreservesNewerManualLastRun guards against a
+// manual TriggerNow's last_run_at being wiped by the scheduler's next tick.
+func TestSetRuntimeStatesSnapshot_PreservesNewerManualLastRun(t *testing.T) {
+	key := "container:project/service"
+
+	runtimeStatesMu.Lock()
+	runtimeStates = map[string]scheduledJobState{}
+	runtimeStatesMu.Unlock()
+
+	manualRun := time.Now()
+	setRuntimeLastRun(key, manualRun)
+
+	// Simulate the loop's next refresh, unaware of the manual run.
+	setRuntimeStatesSnapshot("", map[string]scheduledJobState{
+		key: {nextRun: manualRun.Add(time.Hour)},
+	})
+
+	got := getRuntimeStatesSnapshot()[key]
+	if !got.lastRun.Equal(manualRun) {
+		t.Fatalf("expected manually triggered last run %v to survive scheduler refresh, got %v", manualRun, got.lastRun)
+	}
+
+	// A genuinely newer lastRun must still win.
+	newerRun := manualRun.Add(2 * time.Hour)
+	setRuntimeStatesSnapshot("", map[string]scheduledJobState{
+		key: {lastRun: newerRun},
+	})
+
+	got = getRuntimeStatesSnapshot()[key]
+	if !got.lastRun.Equal(newerRun) {
+		t.Fatalf("expected newer scheduler-tracked last run %v to win, got %v", newerRun, got.lastRun)
+	}
+}
+func TestSetRuntimeStatesSnapshotPreservesOtherContexts(t *testing.T) {
+	runtimeStatesMu.Lock()
+	runtimeStates = map[string]scheduledJobState{
+		"remote::container:shared/job": {deployment: "remote"},
+	}
+	runtimeStatesMu.Unlock()
+
+	setRuntimeStatesSnapshot("", map[string]scheduledJobState{
+		"container:shared/job": {deployment: "default"},
+	})
+
+	states := getRuntimeStatesSnapshot()
+	if len(states) != 2 {
+		t.Fatalf("expected both context partitions, got %#v", states)
+	}
+
+	setRuntimeStatesSnapshot("remote", map[string]scheduledJobState{})
+
+	states = getRuntimeStatesSnapshot()
+	if _, ok := states["remote::container:shared/job"]; ok {
+		t.Fatal("expected stale remote state to be removed")
+	}
+
+	if _, ok := states["container:shared/job"]; !ok {
+		t.Fatal("expected default state to be preserved")
+	}
+}
+func TestClearRuntimeContext(t *testing.T) {
+	t.Cleanup(func() {
+		runtimeStatesMu.Lock()
+		runtimeStates = map[string]scheduledJobState{}
+		runtimeRunStatuses = map[string]string{}
+		runtimeRunningStates = map[string]bool{}
+		runtimeStatesMu.Unlock()
+	})
+
+	runtimeStatesMu.Lock()
+	runtimeStates = map[string]scheduledJobState{
+		"container:shared/job":         {deployment: "default"},
+		"remote::container:shared/job": {deployment: "remote"},
+	}
+	runtimeRunStatuses = map[string]string{
+		"container:shared/job":         "running",
+		"remote::container:shared/job": "exited (0)",
+	}
+	runtimeRunningStates = map[string]bool{
+		"container:shared/job":         true,
+		"remote::container:shared/job": true,
+	}
+	runtimeStatesMu.Unlock()
+
+	clearRuntimeContext("remote")
+
+	if _, ok := getRuntimeStatesSnapshot()["remote::container:shared/job"]; ok {
+		t.Fatal("expected remote runtime state to be removed")
+	}
+
+	if _, ok := getRuntimeRunStatusesSnapshot()["remote::container:shared/job"]; ok {
+		t.Fatal("expected remote run status to be removed")
+	}
+
+	if _, ok := getRuntimeRunningStatesSnapshot()["remote::container:shared/job"]; ok {
+		t.Fatal("expected remote running state to be removed")
+	}
+
+	if _, ok := getRuntimeStatesSnapshot()["container:shared/job"]; !ok {
+		t.Fatal("expected default runtime state to be preserved")
+	}
+}
+func TestUpdateRuntimeRunStatus(t *testing.T) {
+	t.Parallel()
+
+	runtimeStatesMu.Lock()
+	runtimeRunStatuses = map[string]string{}
+	runtimeStatesMu.Unlock()
+
+	job := scheduledJob{
+		key:  "container:project/service",
+		mode: scheduledJobModeContainer,
+	}
+	cfg := docker.JobScheduleConfig{ExecutionMode: docker.JobExecutionModeOneOff}
+
+	updateRuntimeRunStatus(job, cfg, nil)
+
+	if got := getRuntimeRunStatusesSnapshot()[job.key]; got != "exited (0)" {
+		t.Fatalf("updateRuntimeRunStatus() success status=%q want=%q", got, "exited (0)")
+	}
+
+	wrapped := fmt.Errorf("scheduled run: %w", &docker.ContainerExitError{ContainerID: "abc", ExitCode: 143})
+	updateRuntimeRunStatus(job, cfg, wrapped)
+
+	if got := getRuntimeRunStatusesSnapshot()[job.key]; got != "exited (143)" {
+		t.Fatalf("updateRuntimeRunStatus() error status=%q want=%q", got, "exited (143)")
+	}
+}
+func TestJobInfo_ContextField(t *testing.T) {
+	t.Parallel()
+
+	// JobInfo.Context must use the external display value ("default"), not
+	// the internal normalized (empty-string) representation.
+	if got, want := docker.DisplayContextName(""), "default"; got != want {
+		t.Fatalf("docker.DisplayContextName(\"\") = %q, want %q", got, want)
+	}
+}
