+++ internal/scheduler/scheduler_worker_test.go
@@ -1,0 +1,93 @@
+package scheduler
+
+import (
+	"testing"
+	"time"
+
+	"github.com/kimdre/doco-cd/internal/docker"
+)
+
+func TestNextScheduledRun_PreservesScheduleAlignment(t *testing.T) {
+	t.Parallel()
+
+	schedule, err := docker.ParseJobScheduleExpression("@every 1m")
+	if err != nil {
+		t.Fatalf("ParseJobScheduleExpression() failed: %v", err)
+	}
+
+	scheduledAt := time.Date(2026, time.May, 9, 12, 0, 0, 0, time.UTC)
+	now := scheduledAt.Add(250 * time.Millisecond)
+
+	got := nextScheduledRun(schedule, scheduledAt, now)
+
+	want := scheduledAt.Add(time.Minute)
+	if !got.Equal(want) {
+		t.Fatalf("nextScheduledRun() = %s, want %s", got.Format(time.RFC3339Nano), want.Format(time.RFC3339Nano))
+	}
+}
+func TestNextScheduledRun_SkipsMissedRunsWithoutDrift(t *testing.T) {
+	t.Parallel()
+
+	schedule, err := docker.ParseJobScheduleExpression("@every 1m")
+	if err != nil {
+		t.Fatalf("ParseJobScheduleExpression() failed: %v", err)
+	}
+
+	scheduledAt := time.Date(2026, time.May, 9, 12, 0, 0, 0, time.UTC)
+	now := scheduledAt.Add(3*time.Minute + 30*time.Second)
+
+	got := nextScheduledRun(schedule, scheduledAt, now)
+
+	want := time.Date(2026, time.May, 9, 12, 4, 0, 0, time.UTC)
+	if !got.Equal(want) {
+		t.Fatalf("nextScheduledRun() = %s, want %s", got.Format(time.RFC3339Nano), want.Format(time.RFC3339Nano))
+	}
+}
+func TestGetNearestNextRun(t *testing.T) {
+	t.Parallel()
+
+	want := time.Date(2026, time.May, 9, 12, 1, 0, 0, time.UTC)
+
+	got, ok := getNearestNextRun(map[string]scheduledJobState{
+		"later": {
+			nextRun: time.Date(2026, time.May, 9, 12, 5, 0, 0, time.UTC),
+		},
+		"earlier": {
+			nextRun: want,
+		},
+		"zero": {},
+	})
+	if !ok {
+		t.Fatalf("getNearestNextRun() reported no next run")
+	}
+
+	if !got.Equal(want) {
+		t.Fatalf("getNearestNextRun() = %s, want %s", got.Format(time.RFC3339Nano), want.Format(time.RFC3339Nano))
+	}
+}
+func TestParseJobScheduleExpression_NextRunUsesLocalTimezone_Berlin(t *testing.T) {
+	berlin, err := time.LoadLocation("Europe/Berlin")
+	if err != nil {
+		t.Fatalf("time.LoadLocation() failed: %v", err)
+	}
+
+	originalLocal := time.Local
+	time.Local = berlin
+
+	t.Cleanup(func() {
+		time.Local = originalLocal
+	})
+
+	schedule, err := docker.ParseJobScheduleExpression("0 */6 * * *")
+	if err != nil {
+		t.Fatalf("ParseJobScheduleExpression() failed: %v", err)
+	}
+
+	now := time.Date(2026, time.May, 11, 0, 30, 0, 0, time.Local)
+	got := schedule.Next(now)
+	want := time.Date(2026, time.May, 11, 6, 0, 0, 0, time.Local)
+
+	if !got.Equal(want) {
+		t.Fatalf("schedule.Next() = %s, want %s", got.Format(time.RFC3339), want.Format(time.RFC3339))
+	}
+}
